refactor: 설명 부분 분리
