+++ src/components/admin/store/edit-dialog/StoreMobileDescription.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const StoreMobileDescription = () => {
+  return <div></div>;
+};
+
+export default StoreMobileDescription;
