+++ src/components/admin/store/edit-dialog/index.ts
@@ -1,0 +1,1 @@
+export { default as EditDialog } from './EditDialog';
